[MODIFY] - CustomScaffoldBody remove SafeArea wrapper - Documents BookButton - change compatibility in compileOptions in build.gradle file  compileOptions {         sourceCompatibility = JavaVersion.VERSION_1_8         targetCompatibility = JavaVersion.VERSION_1_8     }
