+++ lib/ui/screens/post_route_card.dart
@@ -8,6 +8,7 @@ import 'package:link/ui/utils/expandable_text.dart';
 import 'package:link/ui/widget_extension.dart';
 
 import '../../models/app.dart';
+import '../widgets/buttons/like_icon.dart';
 import '../widgets/photo_view_gallery_widget.dart';
 
 // ignore: must_be_immutable
@@ -369,11 +370,6 @@ class BuildCardFooter extends StatelessWidget {
             isLike: isLike,
             toggleLike: onStarPressed,
           ),
-          /*
-          InkWell(
-              onTap: loading ? null : onStarPressed,
-              child: const Icon(Icons.hotel_class_outlined)),
-              */
         ),
         Expanded(
             child: InkWell(
@@ -394,62 +390,6 @@ class BuildCardFooter extends StatelessWidget {
               child: const Icon(Icons.share)),
         ),
       ],
-    );
-  }
-}
-
-class LikeIcon extends StatefulWidget {
-  const LikeIcon({super.key, this.isLike = false, required this.toggleLike});
-
-  // Initial 'like' state
-  final bool isLike;
-
-  // External function that accepts the updated 'like' state
-  final void Function(bool isLiked) toggleLike;
-
-  @override
-  State<LikeIcon> createState() => _LikeIconState();
-}
-
-class _LikeIconState extends State<LikeIcon> {
-  late final ValueNotifier<bool> _likeNotifier;
-
-  @override
-  void initState() {
-    super.initState();
-    // Initialize ValueNotifier with the initial like state
-    _likeNotifier = ValueNotifier(widget.isLike);
-  }
-
-  // Clean up ValueNotifier to avoid memory leaks
-  @override
-  void dispose() {
-    _likeNotifier.dispose();
-    super.dispose();
-  }
-
-  // Toggle the like state internally and pass the new value to the external callback
-  void _toggleLike() {
-    _likeNotifier.value = !_likeNotifier.value;
-    widget.toggleLike(_likeNotifier.value); // Pass the updated state
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return IconButton(
-      onPressed: _toggleLike, // Call the internal toggle function
-      icon: ValueListenableBuilder<bool>(
-        valueListenable: _likeNotifier,
-        builder: (context, value, child) {
-          return Icon(
-            value ? Icons.recommend : Icons.recommend_sharp,
-            color: value
-                ? Colors.blue
-                : Colors.grey, // Change color based on state
-            size: 20,
-          );
-        },
-      ),
     );
   }
 }
